Fixed a number of issues
This mainly just fixes the block translation issues.
Where blocks are defined in a given version that version is defined and given to the translator rather than using the arbitrary version from the world which may not match the version the blocks were defined in.
Fixed the way block entities are defined. The way I have written it will be forwards compatible when item translation is supported.
Extended some examples
Added loading bar support to some examples
Added a radius option to the sphere operation.
Simplified the chunk location code in the chunk outline operation.